refactor: create BaseTeeAppRunner class
