+++ src/main/kotlin/org/wfanet/measurement/edpaggregator/BUILD.bazel
@@ -36,3 +36,24 @@ kt_jvm_library(
         "@wfa_common_jvm//imports/java/com/google/protobuf/util",
     ],
 )
+
+kt_jvm_library(
+    name = "base_tee_app_runner",
+    srcs = ["BaseTeeAppRunner.kt"],
+    deps = [
+        "//imports/java/com/google/cloud/logging",
+        "//imports/java/com/google/cloud/secretmanager",
+        "//imports/java/com/google/crypto/tink/integration/gcpkms",
+        "//imports/java/io/opentelemetry/instrumentation/grpc",
+        "//src/main/kotlin/org/wfanet/measurement/common/edpaggregator:edp_aggregator_config",
+        "//src/main/kotlin/org/wfanet/measurement/edpaggregator/telemetry",
+        "//src/main/proto/wfa/measurement/config/edpaggregator:event_data_provider_configs_kt_jvm_proto",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto:signing_certs",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto/tink",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/grpc",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/kms:gcloud_to_aws_kms_client_factory",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/kms:gcp_kms_client_factory",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub:google_pub_sub_client",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub:subscriber",
+    ],
+)
