+++ src/main/kotlin/org/wfanet/measurement/edpaggregator/subpoolassigner/BUILD.bazel
@@ -26,17 +26,14 @@ kt_jvm_library(
     srcs = ["SubpoolAssignerAppRunner.kt"],
     deps = [
         ":subpool_assigner_app",
-        "//imports/java/com/google/cloud/logging",
-        "//imports/java/com/google/cloud/secretmanager",
-        "//imports/java/com/google/crypto/tink/integration/gcpkms",
-        "//imports/java/io/opentelemetry/instrumentation/grpc",
-        "//src/main/kotlin/org/wfanet/measurement/common/edpaggregator:edp_aggregator_config",
-        "//src/main/kotlin/org/wfanet/measurement/edpaggregator/telemetry",
-        "//src/main/proto/wfa/measurement/config/edpaggregator:event_data_provider_configs_kt_jvm_proto",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto/tink",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/kms:gcloud_to_aws_kms_client_factory",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/kms:gcp_kms_client_factory",
+        "//src/main/kotlin/org/wfanet/measurement/edpaggregator:base_tee_app_runner",
+        "//src/main/kotlin/org/wfanet/measurement/edpaggregator:storage_config",
+        "//src/main/proto/wfa/measurement/edpaggregator/v1alpha:subpool_assigner_params_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/securecomputation/controlplane/v1alpha:work_item_attempts_service_kt_jvm_grpc_proto",
+        "//src/main/proto/wfa/measurement/securecomputation/controlplane/v1alpha:work_item_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/securecomputation/controlplane/v1alpha:work_items_service_kt_jvm_grpc_proto",
+        "@wfa_common_jvm//imports/java/com/google/crypto/tink",
+        "@wfa_common_jvm//imports/java/com/google/protobuf",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub:google_pub_sub_client",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub:subscriber",
     ],
 )
